params/types/coregeth,params/types/ctypes,params/types/genesisT,params/types/goethereum: install Shanghai=Spiral block number based iface methods
ETC plans to install most EIPs associated with
ETH's Shanghai fork, but ETC doesn't use PoS so activation
by block time, so this adds supporting block-based activation
methods for those features.
Still TODO is to actually implement the associated conditionals
where the Getters are used in the code to actually turn
the new features on.

Date: 2023-08-18 09:11:04-06:00
Signed-off-by: meows <b5c6@protonmail.com>