params/types/goethereum: fix incompatible check; permit passthrough block-based Shanghai config
Failing was TestCheckCompatible, which
converts Geth::CoreGeth configuration types;
CoreGeth's EIP3651 block transition was fatally
inapplicable to Geth's config.

So this patch sidesteps the issue by permitting
the conversion to quietly fail; expecting that users/developers
will be sane enough to configure upgrades with one pattern or the other
but not both.

Date: 2023-08-19 09:17:33-06:00
Signed-off-by: meows <b5c6@protonmail.com>