les: re-install les API methods and service
The absence of these was causing geth
to stall out (RPC server shutting down;
silently because of No-available-methods-for-service
error)

It is reasonable to maintain these
methods because the les service should
still exist because the les feature
still exists.
HOWEVER, currently checkpointing logic
has been removed from the les service,
and all the API methods are related to
checkpointing...
so they all now return the same error
(errNotActivated).

Maybe we'll reconsider removing the checkpointing
from les, but for now
this makes the tests pass, at least.

Date: 2023-10-19 09:12:19-06:00
Signed-off-by: meows <b5c6@protonmail.com>