eth: move comment to harder-to-accidentally-remove line
This point in the code is important.
Its where the txpool is permitted
to accept txes, which is pending
sync completion.
The miner also cares about sync
completion, which is why the fetcher
fires the event here.

This commit only moves the relevvant
comment to a location which is hopefully
harder to accidentally get removed in a merge.

Date: 2023-11-29 16:24:04-07:00
Signed-off-by: meows <b5c6@protonmail.com>